+++ DateConstants.js
@@ -1,6 +1,10 @@
 "use strict";
 var DateConstants;
 (function (DateConstants) {
+    /** String to append to display times that are less than 12 hrs */
+    DateConstants.AM_STRING = "am";
+    /** String to append to display times that are greater than or equal to 12 hrs */
+    DateConstants.PM_STRING = "pm";
     /** Number of milliseconds in a 24 hour period */
     DateConstants.MILLIS_PER_DAY = 1000 * 60 * 60 * 24;
     /** Number of milliseconds in a 7 x 24 hour period */