+++ test/DateTimesTest.js
@@ -8,16 +8,16 @@ suite("DateTimes", function DateTimesTest() {
     function testDotNetJson(asr) {
         var date = new Date(2000, 0, 27);
         var jsonTimestamp = DotNetJsonDate.toDotNetJson(date);
-        asr.equal(date.getTime(), DotNetJsonDate.parseDotNetJson(jsonTimestamp).getTime());
-        asr.equal(new Date(0).getTime(), DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(null)).getTime());
-        asr.equal(new Date(0).getTime(), DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(0)).getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(jsonTimestamp).getTime(), date.getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(null)).getTime(), new Date(0).getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(0)).getTime(), new Date(0).getTime());
     }
     function testDotNetJsonTimestamp(asr) {
         var now = Date.now();
         var jsonTimestamp = Timestamps.toDotNetJson(now);
-        asr.equal(now, Timestamps.parseDotNetJson(jsonTimestamp));
-        asr.equal(new Date(0).getTime(), Timestamps.parseDotNetJson(Timestamps.toDotNetJson(null)));
-        asr.equal(new Date(0).getTime(), Timestamps.parseDotNetJson(Timestamps.toDotNetJson(0)));
+        asr.equal(Timestamps.parseDotNetJson(jsonTimestamp), now);
+        asr.equal(Timestamps.parseDotNetJson(Timestamps.toDotNetJson(null)), new Date(0).getTime());
+        asr.equal(Timestamps.parseDotNetJson(Timestamps.toDotNetJson(0)), new Date(0).getTime());
     }
     test("DateTimes.MILLIS_PER_DAY", function MS_PER_DAYTest() {
         asr.equal(DateConstants.MILLIS_PER_DAY, 24 * 60 * 60 * 1000);
@@ -31,21 +31,21 @@ suite("DateTimes", function DateTimesTest() {
     test("Dates.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
         var res = Dates.getDayMinutes(date);
-        asr.equal(134, res);
+        asr.equal(res, 134);
     });
     test("Dates.toDisplayDate", function toDisplayDateTest() {
         var res1 = Dates.toDisplayDate(new Date(2000, 0, 27, 2, 14));
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
         var res2 = Dates.toDisplayDate(new Date(2000, 0, 9, 2, 14), "-");
-        asr.equal("01-09-2000", res2);
+        asr.equal(res2, "01-09-2000");
     });
     test("Dates.toDisplayDateTime", function toDisplayDateTimeTest() {
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 0, 0), false);
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 2, 14), false);
-        asr.equal("01/27/2000 02:14 a.m.", res1);
+        asr.equal(res1, "01/27/2000 02:14 am");
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 23, 8), false);
-        asr.equal("01/27/2000 11:08 p.m.", res1);
+        asr.equal(res1, "01/27/2000 11:08 pm");
     });
     test("Dates.dayDiff", function dayDiffTest() {
         var oldDate = new Date(2000, 0, 27);
@@ -60,7 +60,7 @@ suite("DateTimes", function DateTimesTest() {
     test("Timestamp.now", function nowTest() {
         var now = new Date().getTime();
         var res1 = Timestamps.now();
-        asr.equal(res1 >= now, true);
+        asr.isTrue(res1 >= now);
     });
     test("Timestamp.toDate", function toDateTest() {
         var now = Timestamps.now();
@@ -76,36 +76,42 @@ suite("DateTimes", function DateTimesTest() {
     test("Timestamp.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
         var res = Timestamps.getDayMinutes(date.getTime());
-        asr.equal(134, res);
+        asr.equal(res, 134);
     });
     test("Timestamp.toDisplayDate", function toDisplayDateTest() {
         var res1 = Timestamps.toDisplayDate(new Date(2000, 0, 27, 2, 14).getTime());
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
         var res2 = Timestamps.toDisplayDate(new Date(2000, 0, 9, 2, 14).getTime(), "-");
-        asr.equal("01-09-2000", res2);
+        asr.equal(res2, "01-09-2000");
     });
     test("Timestamp.toDisplayDateTime", function toDisplayDateTimeTest() {
         var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 0, 0).getTime(), false);
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
         var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 2, 14).getTime(), false);
-        asr.equal("01/27/2000 02:14 a.m.", res1);
+        asr.equal(res1, "01/27/2000 02:14 am");
+        // Testing changing the AM/PM time string constants
+        var origPmStr = DateConstants.PM_STRING;
+        DateConstants.PM_STRING = "P.M.";
+        var res1 = Timestamps.toDisplayDateTime(new Date(1920, 5, 3, 15, 21).getTime(), false);
+        asr.equal(res1, "06/03/1920 03:21 P.M.");
+        DateConstants.PM_STRING = origPmStr;
         var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 23, 8).getTime(), false);
-        asr.equal("01/27/2000 11:08 p.m.", res1);
+        asr.equal(res1, "01/27/2000 11:08 pm");
     });
     test("Timestamp.toDisplayTime", function toDisplayTimeTest() {
         var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 0, 0).getTime());
-        asr.equal("12:00 a.m.", res1);
+        asr.equal(res1, "12:00 am");
         var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 2, 14).getTime());
-        asr.equal("02:14 a.m.", res1);
+        asr.equal(res1, "02:14 am");
         var res1 = Timestamps.toDisplayTime(new Date(2000, 0, 27, 23, 8).getTime());
-        asr.equal("11:08 p.m.", res1);
+        asr.equal(res1, "11:08 pm");
     });
     test("Timestamp.isSameDate", function isSameDateTest() {
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27), new Date(2000, 0, 28));
-        asr.equal(false, res1);
+        asr.isFalse(res1);
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 2, 14), new Date(2000, 0, 27, 23));
-        asr.equal(true, res1);
+        asr.isTrue(res1);
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 23, 8), new Date(2000, 0, 27));
-        asr.equal(true, res1);
+        asr.isTrue(res1);
     });
 });