+++ test/DateTimesTest.ts
@@ -12,20 +12,20 @@ suite("DateTimes", function DateTimesTest() {
     function testDotNetJson(asr: Chai.AssertStatic) {
         var date = new Date(2000, 0, 27);
         var jsonTimestamp = DotNetJsonDate.toDotNetJson(date);
-        asr.equal(date.getTime(), DotNetJsonDate.parseDotNetJson(jsonTimestamp).getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(jsonTimestamp).getTime(), date.getTime());
 
-        asr.equal(new Date(0).getTime(), DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc><any>null)).getTime());
-        asr.equal(new Date(0).getTime(), DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc>0)).getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc><any>null)).getTime(), new Date(0).getTime());
+        asr.equal(DotNetJsonDate.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc>0)).getTime(), new Date(0).getTime());
     }
 
 
     function testDotNetJsonTimestamp(asr: Chai.AssertStatic) {
         var now = <TimestampUtc>Date.now();
         var jsonTimestamp = Timestamps.toDotNetJson(now);
-        asr.equal(now, Timestamps.parseDotNetJson(jsonTimestamp));
+        asr.equal(Timestamps.parseDotNetJson(jsonTimestamp), now);
 
-        asr.equal(new Date(0).getTime(), Timestamps.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc><any>null)));
-        asr.equal(new Date(0).getTime(), Timestamps.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc>0)));
+        asr.equal(Timestamps.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc><any>null)), new Date(0).getTime());
+        asr.equal(Timestamps.parseDotNetJson(Timestamps.toDotNetJson(<TimestampUtc>0)), new Date(0).getTime());
     }
 
 
@@ -47,28 +47,28 @@ suite("DateTimes", function DateTimesTest() {
     test("Dates.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
         var res = Dates.getDayMinutes(date);
-        asr.equal(134, res);
+        asr.equal(res, 134);
     });
 
 
     test("Dates.toDisplayDate", function toDisplayDateTest() {
         var res1 = Dates.toDisplayDate(new Date(2000, 0, 27, 2, 14));
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
 
         var res2 = Dates.toDisplayDate(new Date(2000, 0, 9, 2, 14), "-");
-        asr.equal("01-09-2000", res2);
+        asr.equal(res2, "01-09-2000");
     });
 
 
     test("Dates.toDisplayDateTime", function toDisplayDateTimeTest() {
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 0, 0), false);
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
 
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 2, 14), false);
-        asr.equal("01/27/2000 02:14 a.m.", res1);
+        asr.equal(res1, "01/27/2000 02:14 am");
 
         var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 23, 8), false);
-        asr.equal("01/27/2000 11:08 p.m.", res1);
+        asr.equal(res1, "01/27/2000 11:08 pm");
     });
 
 
@@ -89,7 +89,7 @@ suite("DateTimes", function DateTimesTest() {
     test("Timestamp.now", function nowTest() {
         var now = new Date().getTime();
         var res1 = Timestamps.now();
-        asr.equal(<number><any>res1 >= now, true);
+        asr.isTrue(<number><any>res1 >= now);
     });
 
 
@@ -113,50 +113,61 @@ suite("DateTimes", function DateTimesTest() {
     test("Timestamp.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
         var res = Timestamps.getDayMinutes(<TimestampUtc>date.getTime());
-        asr.equal(134, res);
+        asr.equal(res, 134);
     });
 
 
     test("Timestamp.toDisplayDate", function toDisplayDateTest() {
         var res1 = Timestamps.toDisplayDate(<TimestampUtc>new Date(2000, 0, 27, 2, 14).getTime());
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
 
         var res2 = Timestamps.toDisplayDate(<TimestampUtc>new Date(2000, 0, 9, 2, 14).getTime(), "-");
-        asr.equal("01-09-2000", res2);
+        asr.equal(res2, "01-09-2000");
     });
 
 
     test("Timestamp.toDisplayDateTime", function toDisplayDateTimeTest() {
         var res1 = Timestamps.toDisplayDateTime(<TimestampUtc>new Date(2000, 0, 27, 0, 0).getTime(), false);
-        asr.equal("01/27/2000", res1);
+        asr.equal(res1, "01/27/2000");
 
         var res1 = Timestamps.toDisplayDateTime(<TimestampUtc>new Date(2000, 0, 27, 2, 14).getTime(), false);
-        asr.equal("01/27/2000 02:14 a.m.", res1);
+        asr.equal(res1, "01/27/2000 02:14 am");
+
+        // Testing changing the AM/PM time string constants
+        var origPmStr = DateConstants.PM_STRING;
+        DateConstants.PM_STRING = "P.M.";
+
+        var res1 = Timestamps.toDisplayDateTime(<TimestampUtc>new Date(1920, 5, 3, 15, 21).getTime(), false);
+        asr.equal(res1, "06/03/1920 03:21 P.M.");
+
+        DateConstants.PM_STRING = origPmStr;
 
         var res1 = Timestamps.toDisplayDateTime(<TimestampUtc>new Date(2000, 0, 27, 23, 8).getTime(), false);
-        asr.equal("01/27/2000 11:08 p.m.", res1);
+        asr.equal(res1, "01/27/2000 11:08 pm");
     });
+
 
     test("Timestamp.toDisplayTime", function toDisplayTimeTest() {
         var res1 = Timestamps.toDisplayTime(<TimestampUtc>new Date(2000, 0, 27, 0, 0).getTime());
-        asr.equal("12:00 a.m.", res1);
+        asr.equal(res1, "12:00 am");
 
         var res1 = Timestamps.toDisplayTime(<TimestampUtc>new Date(2000, 0, 27, 2, 14).getTime());
-        asr.equal("02:14 a.m.", res1);
+        asr.equal(res1, "02:14 am");
 
         var res1 = Timestamps.toDisplayTime(<TimestampUtc>new Date(2000, 0, 27, 23, 8).getTime());
-        asr.equal("11:08 p.m.", res1);
+        asr.equal(res1, "11:08 pm");
     });
+
 
     test("Timestamp.isSameDate", function isSameDateTest() {
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27), new Date(2000, 0, 28));
-        asr.equal(false, res1);
+        asr.isFalse(res1);
 
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 2, 14), new Date(2000, 0, 27, 23));
-        asr.equal(true, res1);
+        asr.isTrue(res1);
 
         var res1 = DateUtil.isSameDate(new Date(2000, 0, 27, 23, 8), new Date(2000, 0, 27));
-        asr.equal(true, res1);
+        asr.isTrue(res1);
     });
 
 });